+++ docs/CHANGELOG.md
@@ -39,6 +39,7 @@ All API code that used 'Tag' now references 'Genre' instead
   * 'type' (string) Default: 'Song' ('Song', 'Video', 'Podcast_Episode', 'Channel', 'Broadcast', 'Democratic', 'Live_Stream') //optional
   * 'clear' (integer) 0|1 clear the current playlist on add //optional
 * API::playlist_edit added new parameter 'sort': (0,1) sort the playlist by 'Artist, Album, Song' //optional
+* Api::indexes added parameter 'include': (0,1) include song details with playlists (XML has this by default)
 
 ### Changed
 
@@ -51,6 +52,10 @@ All API code that used 'Tag' now references 'Genre' instead
 * Don't allow duplicate podcast feeds
 * Extend democratic cooldown past 255 and show an error when out of range
 * Api::shares filter is optional
+
+### Deprecated
+
+* Api::indexes will stop including playlist track and id in xml by default in Ampache 5.0.0
 
 ### Fixed